Group commands
It seems prudent to group commands as per `ipfs`
It's easy to lose track of what's important and
relevant when working with IPFS. This should help
developers some.

Add missing functions to Api.

Set specs also in the Api space.

Group commands

It seems prudent to group commands as per `ipfs`
It's easy to lose track of what's important and
relevant when working with IPFS. This should help
developers some.

Add missing functions to Api.

Set specs also in the Api space.