Substantially refactored code structure.
The code is much more maintainable. Most notably
it poretty easy to make exceptions at the single
command, if you need to massage data.

Some struct have been added, so we can have some
integrity with regards to return values. This means
that structs needs to be updated, when upstream their
object structure, but that is considered a good thing.

The old kludgy response handlers are now gone. The new
one doesn't do much but it's a great place to intercept
data and it gives a nice function `handle_api_response` to
be used around the code in an understandable way.

Receiving of raw data through hackneys stream body work well.
The old multipart_add_files function still is around, which
makes it easy to add a whole directory, if needed.
Multipart doesn't really do that, but it should.